requirements.txt
fastapi
uvicorn
gspread
google-auth
dateparser
apscheduler 